+++ IntegrationEngine.Model/IIntegrationJobTrigger.cs
@@ -2,8 +2,10 @@
 
 namespace IntegrationEngine.Model
 {
-    public interface IIntegrationJobTrigger : IHasStringId, IHasStateId
+    public interface IIntegrationJobTrigger : IHasStringId
     {
         string JobType { get; set; }
+        int StateId { get; set; }
+        string StateDescription { get; }
     }
 }