+++ IntegrationEngine/Scheduler/IEngineScheduler.cs
@@ -12,6 +12,7 @@ namespace IntegrationEngine.Scheduler
         void ScheduleJobWithSimpleTrigger(SimpleTrigger triggerDefinition);
         bool IsJobTypeRegistered(string jobTypeName);
         Type GetRegisteredJobTypeByName(string jobTypeName);
+        bool DeleteTrigger(IIntegrationJobTrigger triggerDefinition);
         void Shutdown();
     }
 }