+++ doc/change-log.md
@@ -5,6 +5,30 @@ currentMenu: change-log
 
 # Change Log
 
+## 2.0.0-beta5
+
+__Features__
+* Add health API for indicating server connection status.
+* Remove trigger from scheduler when it is deleted via the API.
+
+__Chores__
+* Move properties from scheduler models into model interfaces.
+
+__Fixes__
+* Cron expression description not generating on Mono.
+
+## 2.0.0-beta4
+
+__Fixes__
+* Use Json.NET to deserialize client responses.
+
+## 2.0.0-beta3
+
+__Features__
+* Implement a typed client library.
+* Add .NET 4 build of client.
+* Add calendar view to dashboard.
+
 ## 2.0.0-beta2
 
 __Chores__